chore: a few more movements
