+++ Genera.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:Build Tools">
+      location = "group:BuildTools">
    </FileRef>
    <FileRef
       location = "group:Utility">
